make new reducer and redux actions files
